+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [3.0.0] - Unreleased
 ### Changed
-- Multiple senders allowed.
+- **-graphiteAddress** takes `hostname:port` pair.
+- Multiple senders allowed by repeating **-graphiteAddress** flag.
+
+### Removed
+- **-graphitePort** flag.
 
 ## [2.2.0] - 2022-06-07
 ### Changed